requirements.txt
numpy==1.11.1
matplotlib==3.2.2
scikit learn==0.22.2
keras==2.6.0
csv==1.0
re==2.2.1
pandas==1.2.1
scipy==1.6.0
pyserial==3.5
re==2.2.1
